Add async_test.cpp
