+++ pilz_testutils/include/pilz_testutils/async_test.h
@@ -17,19 +17,13 @@
 #ifndef ASYNC_TEST_H
 #define ASYNC_TEST_H
 
-#include <ros/ros.h>
-
 #include <mutex>
 #include <condition_variable>
-#include <atomic>
 
 #include <gtest/gtest.h>
 
 namespace testing
 {
-
-#define ACTION_OPEN_BARRIER(str) ::testing::InvokeWithoutArgs([this](void){this->triggerClearEvent(str); return true;})
-#define ACTION_OPEN_BARRIER_VOID(str) ::testing::InvokeWithoutArgs([this](void){this->triggerClearEvent(str);})
 
 /**
  * @brief Test class that allows the handling of asynchronous test objects
@@ -72,40 +66,12 @@ class AsyncTest
     std::set<std::string> waitlist_ {};
 };
 
-void AsyncTest::triggerClearEvent(std::string event)
-{
-  std::lock_guard<std::mutex> lk(m_);
-  if (clear_events_.empty())
-  {
-    waitlist_.insert(event);
-  }
-  else if (clear_events_.erase(event) < 1)
-  {
-    ROS_WARN_STREAM("Triggered event " << event << " despite not waiting for it.");
-  }
-  cv_.notify_one();
-}
-
-void AsyncTest::barricade(std::string clear_event)
-{
-  barricade({clear_event});
-}
-
-void AsyncTest::barricade(std::initializer_list<std::string> clear_events)
-{
-  std::unique_lock<std::mutex> lk(m_);
-  std::copy_if(clear_events.begin(), clear_events.end(), std::inserter(clear_events_, clear_events_.end()),
-               [this](std::string event){ return this->waitlist_.count(event) == 0; });
-  waitlist_.clear();
-  while(!clear_events_.empty())
-  {
-    cv_.wait(lk);
-  }
-}
-
 // for better readability in tests
 #define BARRIER(str) barricade(str)
 #define BARRIER2(str1, str2) barricade(str1, str2)
+
+#define ACTION_OPEN_BARRIER(str) ::testing::InvokeWithoutArgs([this](void){this->triggerClearEvent(str); return true;})
+#define ACTION_OPEN_BARRIER_VOID(str) ::testing::InvokeWithoutArgs([this](void){this->triggerClearEvent(str);})
 }
 
 #endif // ASYNC_TEST_H
